Move both UIActionHandler and IUIActionHandlerProvider to Eutherion.Shared.csproj.
