+++ Sandra.UI/RichTextBoxEx.cs
@@ -19,6 +19,7 @@
 **********************************************************************************/
 #endregion
 
+using Eutherion.UIActions;
 using Eutherion.Win.Controls;
 using Eutherion.Win.UIActions;
