+++ Sandra.UI/SyntaxEditor.cs
@@ -20,6 +20,7 @@
 #endregion
 
 using Eutherion.Text;
+using Eutherion.UIActions;
 using Eutherion.Utils;
 using Eutherion.Win.Storage;
 using Eutherion.Win.UIActions;
