+++ Sandra.UI/UIActionForm.cs
@@ -19,6 +19,7 @@
 **********************************************************************************/
 #endregion
 
+using Eutherion.UIActions;
 using Eutherion.Win;
 using Eutherion.Win.UIActions;
 using System;
